+++ packages/atlas/src/components/_channel/ChannelLink/ChannelLink.tsx
@@ -47,17 +47,14 @@ export const ChannelLink: FC<ChannelLinkProps> = ({
   customTitle,
   followButton = false,
 }) => {
-  const { extendedChannel } = useBasicChannel(id || '', {
+  const { channel } = useBasicChannel(id || '', {
     skip: !id,
     onCompleted: (data) => !data && onNotFound?.(),
     onError: (error) => SentryLogger.error('Failed to fetch channel', 'ChannelLink', error, { channel: { id } }),
   })
-  const { toggleFollowing, isFollowing } = useHandleFollowChannel(
-    extendedChannel?.channel.id,
-    extendedChannel?.channel?.title
-  )
+  const { toggleFollowing, isFollowing } = useHandleFollowChannel(channel?.id, channel?.title)
 
-  const displayedChannel = overrideChannel ? { channel: overrideChannel } : extendedChannel
+  const displayedChannel = overrideChannel ? overrideChannel : channel
 
   const handleFollowButtonClick = (e: MouseEvent) => {
     e.preventDefault()
@@ -73,7 +70,7 @@ export const ChannelLink: FC<ChannelLinkProps> = ({
             withHandle={!hideHandle}
             loading={!displayedChannel}
             size={avatarSize}
-            assetUrls={displayedChannel?.channel.avatarPhoto?.resolvedUrls}
+            assetUrls={displayedChannel?.avatarPhoto?.resolvedUrls}
           />
         </StyledLink>
       )}
@@ -92,12 +89,11 @@ export const ChannelLink: FC<ChannelLinkProps> = ({
                     as="span"
                     color={textSecondary ? 'colorCoreNeutral200' : undefined}
                   >
-                    {customTitle || displayedChannel.channel.title}
+                    {customTitle || displayedChannel.title}
                   </ChannelTitle>
                   {followButton && (
                     <Text as="p" variant="t100" color="colorText" margin={{ top: 1 }}>
-                      {displayedChannel.channel.followsNum}{' '}
-                      {displayedChannel.channel.followsNum === 1 ? 'follower' : 'followers'}
+                      {displayedChannel.followsNum} {displayedChannel.followsNum === 1 ? 'follower' : 'followers'}
                     </Text>
                   )}
                 </StyledLink>
@@ -105,7 +101,7 @@ export const ChannelLink: FC<ChannelLinkProps> = ({
                   <FollowButtonWrapper>
                     <ProtectedActionWrapper
                       title="You want to follow this channel?"
-                      description={`Sign in to follow ${displayedChannel.channel.title}`}
+                      description={`Sign in to follow ${displayedChannel.title}`}
                     >
                       <Button variant="secondary" onClick={handleFollowButtonClick}>
                         {isFollowing ? 'Unfollow' : 'Follow'}